// src/types/api/githubComIntocodePromptarchiveInternalServiceFolderResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * PromptArchive API
 * REST API for storing and managing prompts (Go + Gin + PostgreSQL + Redis)
 * OpenAPI spec version: 1.0
 */

export interface GithubComIntocodePromptarchiveInternalServiceFolderResponse {
  color?: string;
  created_at?: string;
  id?: string;
  name?: string;
  position?: number;
  prompts_count?: number;
  updated_at?: string;
}